Creates and saves model parameter dataframe
With this commit, model_comparison.core.create_model_dataframe will
create the dataframe with all the model paramters and save it as a
csv. Also included are functions for setup, image, and brain data keys,
but these just return a dictionary with those subset of keys because I
think they're less useful.