Adds stimuliNames.mat, gets comparison working
With this commit, the .py and .m scripts will produce predictions that
can be compared against each other.

Next step is to make this comparison a little more structured: set up
script that will run the code on the particular images I'm interested
in (for probably only one voxel), group those predictions in a
meaningful way (using the new stimuliNames.mat) and visualize them.