Starts automation of visualization (WiP)
We would like to automate all the steps up to the visualization shown in
the jupyter notebook, including running the python and matlab code using
the same arguments, etc. I had thought about using a Makefile to do so,
and this is the first steps of that.

This is a work in progress and WILL NOT WORK.