Adds calc_voxel_selector
This commit adds the calc_voxel_selector calculator, an optional step
that limits the model to only looking at the voxels in the list / array
passed to the model as `voxel_idx`.

compare_with_Kay2013.py makes use of this because we're only interested
in a small number of voxels

Additionally, this adds 'pRF_hemispheres' to the calculates wrapper for
calc_pRFs_from_freesurfer_retinotopy. It not being included was
technically a bug but it wasn't necessary before since nothing required
pRF_hemispheres as an argument.