compare_with_Kay2013 now does all frames for a given stimuli
For the stimulus images from Kay2013, some of them had multiple images
per stimulus (they were all roughly equivalent, just variants of each
other). In the paper, predictions were created separately for each image
and then the corresponding ones were averaged together. In order to
allow that, I needed to add a stimulus_names / stimulus_model_names
array to the code (I will probably rename it because right now it's easy
to confuse with stimuli_names, from the .mat that Catherine created),
which is an array created when we create the stimulus images of the form
"####_sub##" where the first #### shows the index into the overall
stimulus array (and so goes from 0 to 260) and the second ## shows the
index into a given stimulus (and so goes from 0 to 9). This array does
not enter the data pool or affect model performance in anyway, but it is
used by model_comparison/core.py, which uses it to name the
image-related columns of the dataframe. Similarly, compareWithKay2013.m
uses it to determine which columns it should grab parameters from and
how it should name the output columns.

In order to make this work, there are also changes in the visualization
notebook. We now show each voxel separately (using factorplot),
averaging across subimages