adds more to Makefile, makes compare_with_Kay2013 options standard
Makefile onw includes a cleantmps and fullclean, which try to clean
everything up (except stimuli.mat, since that takes a while to download
and isn't otherwise used)

the options used for compare_with_Kay2013 are now be good options that
provide the right results: the matlab and python now look qualitatively
similar.

Now, need to move visualization into a script and make that part of the
makefile.