Bugfix: matlab pRF uses same part of image as py
With this bugfix, the matlab pRFs now capture the same location / area
in the image as the python. The matlab version of the code ends up
operating on images that are 90x90 as opposed to 180x180 (the call to
applymultiscalegaborfilters shrinks them). Therefore, we need to divide
the pRF center and size by 2 to make sure they refer to the same part of
the image.