Adds several more visualization options
Biggest change here is how we handle condition: _create_plot_df doesn't
/only/ include those images that correspond to the condition. Instead,
we have a new column, image_name, which contains the information shown
in stimuli_descriptions, which allows easy later filtering.

This means that _plot_stimuli had to change slightly so that it makes
use of stimuli_descriptions when plotting a condition (if that condition
is a string).

This commit also adds several more options to the visualizations:

 - the sns.factorplot now makes use of order so that, if the conditions
   were given as a list / array of indices, we preserve that order

 - users can now specify plot_kwargs to set various components of
   sns.factorplot themselves. we have default options but only use them
   if the corresponding key is not already in plot_kwargs

 - adds three more visualizations to the default call: 'horizontal
   sweep', 'vertical sweep', and 'full'. Full includes more of the
   stimuli, all the ones in figure 5 of the Kay2013 paper plus size, for
   easy summarizing. horizontal and vertical sweep look at the
   corresponding horizontal and vertical sweeps of the squiggly lines.