compare_with_Kay2013 now uses **kwargs
With this, the compare_with_Kay2013 function now uses **kwargs to allow
the setting of some of the parameters from an outside call. The function
will set defaults if some keys are missing from kwargs, but it won't
overwrite them. I need to check if this is the correct way to do it, but
this makes it more flexible.

Also, small changes to freq_pref_check:

 - filenames now use %02d formatting (so they sort the way we want)

 - now uses 1 cpd instead of 3

 - checks 1 through 21 cycles per image

 - renames frequency to cycles per image in plot_df

 - returns plot_df