Cleans up, separates batch and notebook workflows
With this (which needs more testing), I reverted the main,
check_pref_across_frequencies, check_pref_across_degrees_per_pixel, and
check_response_across_prefs functions to before so they each work
independently (and should be called from a notebook). Adds new
functions, check_full_space and combine_dfs, that tries to explore the
full space, looking at all combinations of frequency preference, spatial
frequencies in images, and stimulus pixels per degree. It's meant to be
run in batches, submitted to the cluster or something, because it takes
a fair amount of time.