Bugfix: ensure gabor filters have mean 0
There was a little bug where the real component of the gabor filters
returned by skimage was not balanced; sometimes the mean was slightly
off. This was a small effect, but we want them to have mean 0. We then
have to divide by the sum of the squares of the real and imaginary parts
to ensure that the power is 1