Bugfix: ensure stimulus_images lie between 0 and 1
Lots of the code assumes that all values in stimulus_images lie between
0 and 1 and we were relying on skimage.util.img_as_float to take care of
this. However, if the image passed to img_as_float already has dtype
np.float64, then the function returns the image unchanged. In order to
correct this, we assume the image lies between 0 and 255 unless the user
sets img_already_rescaled (a new optional paramter) to True. When
img_already_rescaled is False, we divide all the images by `255.` to
rescale.

Other changes in this commit:

 - adds img_already_rescaled to README

 - compare_with_Kay2013 now uses the default stimulus chain, because
   it's able to handle stimulus_images being set directly