Large changes to Metamers.ipynb (WiP)
With this, we're working towards finalizing the notebook, though it's
still not done. Major additions:

 - starts on explanatory text

 - code to create SNR dataframes

 - code to run full and "dumb V1" versions of the model

 - visualizations of predicted responses and SNR

 - visualizations of predicted responses projected onto to the
   image ("cortical images"), though this needs refinement

 - starts visualizations of predicted responses projected onto the
   cortex, both inflated (using pysurfer) and flattened (using custom
   code), though this requires more work.

eventually, some of this code, especially the various visualizations,
will end up in other places in the SCO repository. additionally, will
write a script to make use of the cluster, since running the model can
take some time.