Adds code to create ImageNet metamers
I don't include code to automatically download the base images, but this
will process them. The important new step is to make sure they're the
right size (that is, can be divided by 2 five times -- this probably
changes with different NUM_SCALES and NUM_ORIENTATIONS, but we won't
worry about that for now).