Fixed bootstrap calculation
I was bootstrapping incorrectly. Before bootstrapping, we want to
average values so that each (voxel, image_name) pair has one value,
which we then resample a number of times equal to the number of unique
image_name values. (we make it possible to change image_name for another
column, though the documentation isn't clear)